# reservations/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations API",
        "description": "Sistema de reservas de mesas con asignación automática de ubicación y combinación inteligente de mesas",
        "contact": {
            "name": "Federico",
            "email": "federico@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear nueva reserva",
                "description": "Crea una reserva con asignación automática de ubicación (A→B→C→D) y combinación óptima de hasta 3 mesas",
                "operationId": "1aabc7d432cc652e63cfc3fb514132cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "reservation_date",
                                    "reservation_time",
                                    "party_size"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID del usuario",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "description": "Fecha de la reserva (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-24"
                                    },
                                    "reservation_time": {
                                        "description": "Hora de la reserva (HH:mm). L-V: 10:00-24:00, Sáb: 22:00-02:00, Dom: 12:00-16:00",
                                        "type": "string",
                                        "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$",
                                        "example": "19:00"
                                    },
                                    "party_size": {
                                        "description": "Cantidad de personas (máximo 18: 3 mesas × 6 personas)",
                                        "type": "integer",
                                        "maximum": 18,
                                        "minimum": 1,
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva creada exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "reservation_date": {
                                                    "type": "string",
                                                    "example": "2025-12-24"
                                                },
                                                "reservation_time": {
                                                    "type": "string",
                                                    "example": "19:00"
                                                },
                                                "party_size": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "location": {
                                                    "description": "Ubicación asignada automáticamente",
                                                    "type": "string",
                                                    "example": "A"
                                                },
                                                "duration_minutes": {
                                                    "description": "Duración por defecto: 2 horas",
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "tables": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "location": {
                                                                "type": "string",
                                                                "example": "A"
                                                            },
                                                            "table_number": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "capacity": {
                                                                "type": "integer",
                                                                "example": 4
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación o sin disponibilidad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No hay mesas disponibles para la fecha y hora solicitada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reservations/{id}/cancel": {
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancelar reserva",
                "description": "Cancela una reserva existente. Solo se pueden cancelar reservas futuras que no estén ya canceladas",
                "operationId": "04d6d9eef11740b453ebda17711b7f55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva a cancelar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva cancelada exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "cancelled"
                                                },
                                                "reservation_date": {
                                                    "type": "string",
                                                    "example": "2025-12-24"
                                                },
                                                "reservation_time": {
                                                    "type": "string",
                                                    "example": "19:00"
                                                },
                                                "party_size": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "A"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al cancelar (ya cancelada o reserva pasada)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La reserva ya está cancelada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}